+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # bonus-server-php-sdk change log
 
-## 0.5.0 (24.07.2018)
-- [issue#36](https://github.com/rarus/bonus-server-php-sdk/issues/36) подавить ошибочный 404 статус если нет скидок, результаты рассчёта скидок стали опциональными
-- [issue#31](https://github.com/rarus/bonus-server-php-sdk/issues/31) добавлен метод `getByBarcode` реализующий получение карты по штрихкоду для роли `Organization` 
+## 0.5.0 (26.07.2018)
+- исправлен ошибочный 404 статус если нет скидок, результаты рассчёта скидок стали опциональными [issue#36](https://github.com/rarus/bonus-server-php-sdk/issues/36)
+- для сущности `Cards` для роли `Organization` добавлен метод `getByBarcode` [issue#31](https://github.com/rarus/bonus-server-php-sdk/issues/31) 
+- для сущности `Cards` для роли `Organization` добавлен метод `getByUser` [issue#37](https://github.com/rarus/bonus-server-php-sdk/issues/37)
+- добавлен объект постраничной навигации `Pagination`
 
 ## 0.3.5 (24.07.2018)
 - для сущности `Transactions` в транспорте для роли `Organization` добавлен метод `addRefundTransaction`